Rename folder & add tests
- Rename folder of create-named-reference.component.html

- Add test for create-named-reference.component.spec.ts
